Fix bugs in name grid stopping editing of display
The old selected value recorded the old, old selected value. This has been fixed.
The selected value ended up in a strange state when editing finished, where it looked like a selection
had occurred but it really hadnt. This is fixed by using the selected index to set the value.